requirements.txt
wget==3.2
tensorflow-gpu==1.0.0
scipy==0.19.1
scikit-learn==0.19.0
tqdm==4.10.0
matplotlib==2.0.2
numpy==1.13.1
six==1.10.0